+++ index.html
@@ -406,7 +406,7 @@
     crossorigin=""
     nonce="abc123"
   ></script>
-  <script src="app.js" nonce="abc123"></script>
+  <script type="module" src="./src/main.js" nonce="abc123"></script>
   <script src="tests.js" nonce="abc123"></script>
   <button id="run-tests-btn" style="position: fixed; bottom: 10px; right: 10px; z-index: 10000; padding: 10px; background-color: lightgoldenrodyellow; border: 1px solid orange;">Run Tests</button>
 </body>
