+++ src/main.js
@@ -1,54 +1,9 @@
+import { populateOrderNumber } from './order.js';
+import { initMenuToggles, initDraggableModals, onDragStart, onDragEnd } from './modal.js';
+import { initMap } from './map.js';
+
 // Basic app logic
-
 document.addEventListener('DOMContentLoaded', () => {
-
-  // --- Global Drag State Variables ---
-  let activeModal = null;
-  let offsetX = 0, offsetY = 0;
-
-  // --- Order Number Generation ---
-  // Attempt to load lastOrderDate and orderSequence from sessionStorage.
-  // Initialize to defaults if not found.
-  // Note: This client-side approach to order number generation has limitations.
-  // A robust, unique order number generation system typically requires a backend component
-  // to ensure uniqueness across all clients and sessions, and to prevent tampering.
-  let lastOrderDate = sessionStorage.getItem('lastOrderDate') || '';
-  let orderSequence = parseInt(sessionStorage.getItem('orderSequence')) || 1;
-
-  function generateOrderNumber() {
-    const now = new Date();
-    const year = now.getFullYear();
-    const month = String(now.getMonth() + 1).padStart(2, '0');
-    const day = String(now.getDate()).padStart(2, '0');
-    const hours = String(now.getHours()).padStart(2, '0');
-    const minutes = String(now.getMinutes()).padStart(2, '0');
-    const seconds = String(now.getSeconds()).padStart(2, '0');
-
-    const currentDate = `${year}${month}${day}`;
-
-    if (currentDate !== lastOrderDate) {
-      orderSequence = 1; // Reset sequence if day has changed
-      lastOrderDate = currentDate;
-    }
-
-    const sequenceStr = String(orderSequence).padStart(3, '0');
-    orderSequence++;
-
-    // Save the updated values to sessionStorage
-    sessionStorage.setItem('lastOrderDate', lastOrderDate);
-    sessionStorage.setItem('orderSequence', orderSequence.toString());
-
-    return `${currentDate}-${hours}${minutes}${seconds}-${sequenceStr}`;
-  }
-
-  function populateOrderNumber() {
-    const orderNumberField = document.getElementById('order-number');
-    if (orderNumberField) {
-      orderNumberField.value = generateOrderNumber();
-    } else {
-      console.error('Error: Order number field not found.');
-    }
-  }
 
   // --- Generic Dynamic Field Helper Functions ---
 
@@ -650,185 +605,3 @@ document.addEventListener('DOMContentLoaded', () => {
   }
 });
 
-// Function to initialize menu toggle buttons
-function initMenuToggles() {
-  const clientMenuToggleBtn = document.getElementById('vertical-client-toggle');
-  const driverMenuToggleBtn = document.getElementById('vertical-driver-toggle');
-  const clientMenuContainer = document.getElementById('client-menu-container'); // Renamed for clarity
-  const driverMenuContainer = document.getElementById('driver-menu-container'); // Renamed for clarity
-  const closeClientMenuBtn = document.getElementById('close-client-menu-btn');
-  const closeDriverMenuBtn = document.getElementById('close-driver-menu-btn');
-
-  if (!clientMenuToggleBtn || !driverMenuToggleBtn || !clientMenuContainer || !driverMenuContainer || !closeClientMenuBtn || !closeDriverMenuBtn) {
-    console.error('Error: One or more menu elements (toggle buttons, containers, or close buttons) not found. Ensure all IDs are correct.');
-    return;
-  }
-
-  function resetModalPosition(modalContentElement) {
-    if (modalContentElement) {
-      modalContentElement.style.left = '';
-      modalContentElement.style.top = '';
-      modalContentElement.style.transform = ''; // Reset transform for CSS centering
-    }
-  }
-
-  function hideClientMenu() {
-    if (clientMenuContainer) {
-      clientMenuContainer.style.display = 'none';
-      clientMenuContainer.hidden = true;
-      clientMenuToggleBtn.setAttribute('aria-expanded', 'false');
-    }
-  }
-
-  function hideDriverMenu() {
-    if (driverMenuContainer) {
-      driverMenuContainer.style.display = 'none';
-      driverMenuContainer.hidden = true;
-      driverMenuToggleBtn.setAttribute('aria-expanded', 'false');
-    }
-  }
-
-  clientMenuToggleBtn.addEventListener('click', () => {
-    const isClientMenuVisible = clientMenuContainer.style.display === 'block';
-    
-    if (isClientMenuVisible) {
-      hideClientMenu();
-    } else {
-      const modalContent = clientMenuContainer.querySelector('.menu-modal-content');
-      resetModalPosition(modalContent);
-      clientMenuContainer.style.display = 'block';
-      clientMenuContainer.hidden = false;
-      clientMenuToggleBtn.setAttribute('aria-expanded', 'true');
-      hideDriverMenu();
-    }
-  });
-
-  driverMenuToggleBtn.addEventListener('click', () => {
-    const isDriverMenuVisible = driverMenuContainer.style.display === 'block';
-
-    if (isDriverMenuVisible) {
-      hideDriverMenu();
-    } else {
-      const modalContent = driverMenuContainer.querySelector('.menu-modal-content');
-      resetModalPosition(modalContent);
-      driverMenuContainer.style.display = 'block';
-      driverMenuContainer.hidden = false;
-      driverMenuToggleBtn.setAttribute('aria-expanded', 'true');
-      hideClientMenu();
-    }
-  });
-
-  // Event listeners for close buttons
-  closeClientMenuBtn.addEventListener('click', hideClientMenu);
-  closeDriverMenuBtn.addEventListener('click', hideDriverMenu);
-
-  // Add close on overlay click
-  if (clientMenuContainer) {
-    clientMenuContainer.addEventListener('click', function(event) {
-      // event.target is the #client-menu-container (the overlay)
-      // NOT clientMenuContainer.querySelector('.menu-modal-content')
-      if (event.target === clientMenuContainer) {
-        hideClientMenu();
-      }
-    });
-  }
-
-  if (driverMenuContainer) {
-    driverMenuContainer.addEventListener('click', function(event) {
-      if (event.target === driverMenuContainer) {
-        hideDriverMenu();
-      }
-    });
-  }
-}
-
-// --- Modal Dragging Functions ---
-function onDragStart(event) {
-  const header = event.target.closest('.modal-header');
-  if (!header) return;
-
-  activeModal = header.closest('.menu-modal-content');
-  if (!activeModal) return;
-
-  // Allow dragging on all screen sizes
-
-  event.preventDefault(); // Prevent text selection, etc., *only if dragging*
-
-  // Ensure transform is cleared so left/top work as absolute pixel values for dragging
-  // The initial centering is done by CSS transform. Once dragged, we switch to top/left.
-  activeModal.style.transform = 'none';
-
-  offsetX = event.clientX - activeModal.getBoundingClientRect().left;
-  offsetY = event.clientY - activeModal.getBoundingClientRect().top;
-
-  document.addEventListener('mousemove', onDrag);
-  document.addEventListener('mouseup', onDragEnd);
-  activeModal.classList.add('modal-dragging');
-}
-
-function onDrag(event) {
-  if (!activeModal) return;
-
-  let newLeft = event.clientX - offsetX;
-  let newTop = event.clientY - offsetY;
-
-  const header = activeModal.querySelector('.modal-header'); // For boundary check
-  if(header){
-    // Basic boundary check to keep modal header roughly in view
-    newLeft = Math.max(-activeModal.offsetWidth + header.offsetWidth, Math.min(newLeft, window.innerWidth - header.offsetWidth));
-    newTop = Math.max(0, Math.min(newTop, window.innerHeight - header.offsetHeight));
-  }
-
-
-  activeModal.style.left = newLeft + 'px';
-  activeModal.style.top = newTop + 'px';
-}
-
-function onDragEnd() {
-  if (!activeModal) return;
-
-  document.removeEventListener('mousemove', onDrag);
-  document.removeEventListener('mouseup', onDragEnd);
-  activeModal.classList.remove('modal-dragging');
-  activeModal = null;
-}
-
-function initDraggableModals() {
-  const modalHeaders = document.querySelectorAll('.modal-header');
-  modalHeaders.forEach(header => {
-    header.addEventListener('mousedown', onDragStart);
-  });
-}
-
-// Function to initialize the Leaflet map
-function initMap() {
-  const mapPlaceholder = document.getElementById('map-placeholder');
-
-  if (!mapPlaceholder) {
-    console.error('Error: Map placeholder element not found.');
-    return;
-  }
-
-  // Check if map is already initialized
-  if (mapPlaceholder._leaflet_id) {
-    console.warn('Warning: Map already initialized.');
-    return;
-  }
-
-  try {
-    const map = L.map('map-placeholder').setView([51.505, -0.09], 13);
-
-    L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
-      attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-    }).addTo(map);
-
-    L.marker([51.505, -0.09]).addTo(map)
-      .bindPopup('A default marker at [51.505, -0.09].')
-      .openPopup();
-      
-    console.log('Map initialized successfully.');
-  } catch (error) {
-    console.error('Error: Error initializing Leaflet map: ' + error.message);
-    mapPlaceholder.innerHTML = '<p style="color:red; text-align:center; font-weight:bold;">Map Error: Could not initialize the map. ' + error.message + '</p>';
-  }
-}
